% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQTpower.R
\name{EQT.power}
\alias{EQT.power}
\title{EqT Power Calculator}
\usage{
EQT.power(
  r,
  low.sesoi,
  up.sesoi,
  n = NULL,
  power = NULL,
  int.size = 0.9,
  n.sims = NULL
)
}
\arguments{
\item{r}{is the predicted Pearson's r correlation (must be between -1 and 1).}

\item{low.sesoi}{is the lower smallest effect in Pearson's r correlation (must be between -1 and 1 and must be less than r).}

\item{up.sesoi}{is the upper smallest effect in Pearson's r correlation (must be between -1 and 1 and must be greater than r).}

\item{n}{is the sample size (# of pairs). Leave as NULL to calculate the necessary sample size with a given power.}

\item{power}{is the desired power (must be between 0 and 1).}

\item{int.size}{is the size of the two-tailed confidence interval for EqT (must be between 0 and 1). Default is .9.}

\item{n.sims}{is the number of simulations used if calculating power from a given n. Defaults to 1000. n.sims is not relevant in calculating necessary sample size. Increase n.sims for greater accuracy.}
}
\description{
This function helps to conduct power analysis for Equivalence Testing with correlational data. To determine the necessary sample size, leave n=NULL and designate the desired power (i.e. power = .8). To calculate the power with a given sample size, leave power=NULL and specify a sample size (i.e. n = 150).
}
\examples{
EQT.power(r = 0, low.sesoi = -.2, up.sesoi = .2, n = NULL, power = NULL, int.size = .95)
}
\keyword{Equivalence}
