% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MESTpower.R
\name{MEST.power}
\alias{MEST.power}
\title{MEST Power Calculator}
\usage{
MEST.power(r, H0, n = NULL, power = NULL, int.size = 0.9, n.sims = NULL)
}
\arguments{
\item{r}{is the predicted Pearson's r correlation (must be between -1 and 1).}

\item{H0}{is the null hypothesis (H0'), which corresponds to the Greatest Effect of No Interest (GENI) in Pearson's r correlation (must be between -1 and 1).}

\item{n}{is the sample size (# of pairs). Leave as NULL to calculate the necessary sample size with a given power.}

\item{power}{is the statistical power (must be between 0 and 1). Leave as NULL to calculate the simulated power with a given sample size.}

\item{int.size}{is the size of the two-tailed confidence interval for MEST (must be between 0 and 1). Default is .9.}

\item{n.sims}{is the number of simulations used if calculating power from a given n. Defaults to 1000. n.sims is not relevant in calculating necessary sample size. Increase n.sims for greater accuracy.}
}
\description{
This function helps to conduct power analysis for MEST with correlational data. To determine the necessary sample size, leave n=NULL and designate a designated power. To calculate the power with a given sample size, leave power=NULL and specify a sample size (n). \cr\cr
See the article by Smiley, Glazier, and Shoda for details on how to conduct MEST: https://psyarxiv.com/465ha/
}
\examples{
## Calculate necessary sample size:
MEST.power(r = .4, H0 = .2, n = NULL, power = .85, int.size = .95)

## Calculate statistical power:
MEST.power(r = .4, H0 = .2, n = 100, power = NULL, int.size = .95, n.sims = 10000)
}
\keyword{MEST}
